Revert "feat(workbench): implement useFetchFlows hook and associated tests (#312)"
This reverts commit 264c0c62ba95246bb2f957edbb59feac76ef3357.
