+++ packages/core/src/__tests__/flow-endpoint.test.ts
@@ -1,10 +1,10 @@
 import { createApiStep, createEventStep } from './fixtures/step-fixtures'
+import { generateFlow } from '../flows-endpoint'
 import { LockedData } from '../locked-data'
 import { z } from 'zod'
 import { Step } from '../types'
 import zodToJsonSchema from 'zod-to-json-schema'
 import { NoPrinter } from '../printer'
-import { generateFlow } from '../helper/flows-helper'
 
 const mockFlowSteps: Step[] = [
   createApiStep({
