+++ packages/core/src/streams/flows-stream.ts
@@ -1,37 +1,37 @@
 import { LockedData } from '../locked-data'
 import { StreamAdapter } from './adapters/stream-adapter'
-import { generateFlow } from '../helper/flows-helper'
-import { FlowResponse } from '../types/flows-types'
 
-export class FlowsStream extends StreamAdapter<FlowResponse> {
+export type Flow = {
+  id: string
+  name: string
+}
+
+export class FlowsStream extends StreamAdapter<Flow> {
   constructor(private readonly lockedData: LockedData) {
     super()
   }
 
-  async get(_: string, id: string): Promise<FlowResponse | null> {
-    const flow = this.lockedData.flows[id]
-
-    if (!flow) {
-      return null
-    }
-
-    return generateFlow(id, flow.steps)
+  async get(id: string): Promise<Flow | null> {
+    return (
+      Object.entries(this.lockedData.flows)
+        .map(([id, flow]) => ({ id, name: flow.name }))
+        .find((flow) => flow.id === id) ?? null
+    )
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  async delete(_: string): Promise<FlowResponse | null> {
-    return null
+  async delete(id: string): Promise<Flow | null> {
+    return { id, name: id }
   }
 
-  async set(_: string, __: string, data: FlowResponse): Promise<FlowResponse> {
+  async set(_: string, __: string, data: Flow): Promise<Flow> {
     return data
   }
 
-  async getGroup(): Promise<FlowResponse[]> {
+  async getGroup(): Promise<Flow[]> {
     /**
      * Get list should receive a groupId argument but that's irrelevant for this stream
      * since we only have one group of flows.
      */
-    return Object.entries(this.lockedData.flows).map(([id, flow]) => generateFlow(id, flow.steps))
+    return Object.entries(this.lockedData.flows).map(([id, flow]) => ({ id, name: flow.name }))
   }
 }
