--- packages/core/src/types/flows-config-types.ts
@@ -1,9 +1,0 @@
-export interface FlowConfig {
-  id: string
-  config: {
-    [filePath: string]: {
-      x: number
-      y: number
-    }
-  }
-}
