+++ packages/stream-client-react/src/motia-stream-provider.tsx
@@ -25,5 +25,7 @@ export const MotiaStreamProvider: React.FC<Props> = ({ children, address }) => {
     return () => stream.close()
   }, [address])
 
+  if (!stream) return null
+
   return <MotiaStreamContext.Provider value={{ stream }}>{children}</MotiaStreamContext.Provider>
 }
