+++ packages/workbench/jest.config.js
@@ -15,43 +15,10 @@ module.exports = {
           jsx: 'react-jsx',
           esModuleInterop: true,
           allowSyntheticDefaultImports: true,
-          types: ['jest', '@testing-library/jest-dom', 'node'],
-          module: 'commonjs',
-          target: 'es2020',
-          lib: ['es2020', 'dom', 'dom.iterable'],
-          moduleResolution: 'node',
-          strict: true,
-          skipLibCheck: true,
-          isolatedModules: true,
+          types: ['jest', '@testing-library/jest-dom'],
         },
-        isolatedModules: true,
-        useESM: false,
-      },
-    ],
-    '^.+\\.(js|jsx)$': [
-      'ts-jest',
-      {
-        tsconfig: {
-          allowJs: true,
-          jsx: 'react-jsx',
-          esModuleInterop: true,
-          allowSyntheticDefaultImports: true,
-          module: 'commonjs',
-          target: 'es2020',
-        },
-        isolatedModules: true,
-        useESM: false,
       },
     ],
   },
-  transformIgnorePatterns: [
-    'node_modules/(?!(@motiadev/stream-client-react|@motiadev/stream-client-browser)/)',
-  ],
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
-  testEnvironmentOptions: {
-    url: 'http://localhost',
-  },
-  clearMocks: true,
-  resetMocks: true,
-  restoreMocks: true,
 }
