+++ packages/workbench/package.json
@@ -32,7 +32,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "dagre": "^0.8.5",
-    "fast-deep-equal": "^3.1.3",
     "date-fns": "^4.1.0",
     "json-schema": "^0.4.0",
     "lucide-react": "^0.510.0",
@@ -50,16 +49,16 @@
     "zod": "^3.24.4"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.1.0",
-    "@testing-library/user-event": "^14.5.2",
     "@types/dagre": "^0.7.52",
     "@types/express": "^5.0.1",
-    "@types/jest": "^29.5.14",
     "@types/json-schema": "^7.0.15",
     "@types/node": "^22.15.18",
     "@types/react": "^19.1.4",
     "@types/react-dom": "^19.1.5",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "ts-jest": "^29.3.4"
