+++ packages/workbench/src/hooks/use-list-flows.tsx
@@ -1,10 +1,8 @@
 import { useStreamGroup } from '@motiadev/stream-client-react'
-import { FlowResponse } from '@/views/flow/hooks/use-get-flow-state'
 
-export type Flow = {
+type Flow = {
   id: string
   name: string
-  data?: FlowResponse
 }
 
 export const useListFlows = () => {
