+++ packages/workbench/src/routes/flow.tsx
@@ -1,11 +1,24 @@
 import { FlowView } from '@/views/flow/flow-view'
+import { FlowConfigResponse, FlowResponse } from '@/views/flow/hooks/use-get-flow-state'
+import { useCallback, useEffect, useState } from 'react'
 import { useParams } from 'react-router'
-import { useFetchFlows } from '@/hooks/use-fetch-flows'
 
 export const Flow = () => {
   const { id } = useParams()
+  const flowId = id!
+  const [flow, setFlow] = useState<FlowResponse | null>(null)
+  const [flowConfig, setFlowConfig] = useState<FlowConfigResponse | null>(null)
 
-  const { flow, flowConfig } = useFetchFlows(id!)
+  const fetchFlow = useCallback(() => {
+    Promise.all([fetch(`/flows/${flowId}`), fetch(`/flows/${flowId}/config`)])
+      .then(([flowRes, configRes]) => Promise.all([flowRes.json(), configRes.json()]))
+      .then(([flow, config]) => {
+        setFlow(flow)
+        setFlowConfig(config)
+      })
+  }, [flowId])
+
+  useEffect(fetchFlow, [fetchFlow])
 
   if (!flow || flow.error)
     return (
