+++ packages/workbench/src/views/flow/hooks/use-get-flow-state.tsx
@@ -32,10 +32,7 @@ export type FlowResponse = {
 }
 
 export type FlowConfigResponse = {
-  id: string
-  config: {
-    [stepName: string]: Position
-  }
+  [key: string]: Position
 }
 
 type FlowEdge = {
@@ -51,7 +48,7 @@ type Position = {
 }
 
 const getNodePosition = (flowConfig: FlowConfigResponse, stepName: string): Position => {
-  const position = flowConfig?.config[stepName]
+  const position = flowConfig[stepName]
   return position || { x: 0, y: 0 }
 }
